requirements.txt
# ──────────────────────────────────────────────────────────────────────
# requirements.txt
# Google Colab đã có sẵn: torch, torchvision, numpy, PIL, matplotlib
# Chỉ cần cài thêm các thư viện bên dưới
# ──────────────────────────────────────────────────────────────────────

gdown>=4.7.1
torchinfo>=1.8.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
pyyaml>=6.0